+++ pants-plugins/oci/CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.5.0 - 2023-11-19
 
 - Fix a crash when image pulling fails
+- Add `nightly` version for `umoci` with experimental support for Mac
+- Add `v1.13.3` version for `skopeo` with experimental support for Mac
 
 ## 0.4.0 - 2023-06-18